requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
